Refactoring to address comments.
