+++ romancal/outlier_detection/outlier_detection_step.py
@@ -56,93 +56,88 @@ class OutlierDetectionStep(RomanStep):
         self.input_container = isinstance(self.input_models, ModelContainer)
         self.skip = False
 
-        # make sure input is a ModelContainer
-        if self.input_container:
-            return self._run_process()
+        # validation
+        if self.input_container and (
+            len(self.input_models) >= 2
+            and all(
+                model.meta.exposure.type == "WFI_IMAGE" for model in self.input_models
+            )
+        ):
+            # Setup output path naming if associations are involved.
+            asn_id = self.input_models.asn_table.get("asn_id", None)
+            if asn_id is not None:
+                _make_output_path = self.search_attr(
+                    "_make_output_path", parent_first=True
+                )
+                self._make_output_path = partial(_make_output_path, asn_id=asn_id)
 
-        # if input is not a ModelContainer, log warning message and raise TypeError
-        self.log.warning(
-            "Input is not a ModelContainer. Outlier detection step will be SKIPPED."
-        )
-        raise TypeError
+            detection_step = outlier_detection.OutlierDetection
+            pars = {
+                "weight_type": self.weight_type,
+                "pixfrac": self.pixfrac,
+                "kernel": self.kernel,
+                "fillval": self.fillval,
+                "nlow": self.nlow,
+                "nhigh": self.nhigh,
+                "maskpt": self.maskpt,
+                "grow": self.grow,
+                "snr": self.snr,
+                "scale": self.scale,
+                "backg": self.backg,
+                "kernel_size": self.kernel_size,
+                "save_intermediate_results": self.save_intermediate_results,
+                "resample_data": self.resample_data,
+                "good_bits": self.good_bits,
+                "allowed_memory": self.allowed_memory,
+                "in_memory": self.in_memory,
+                "make_output_path": self.make_output_path,
+                "resample_suffix": "i2d",
+            }
 
-    def _run_process(self):
-        # check conditions for valid input
-        self.valid_input = len(self.input_models) >= 2 and all(
-            model.meta.exposure.type == "WFI_IMAGE" for model in self.input_models
-        )
+            self.log.debug(
+                f"Using {detection_step.__name__} class for outlier_detection"
+            )
 
-        if self.valid_input:
-            return self._run_outlier_detection()
+            # Set up outlier detection, then do detection
+            step = detection_step(self.input_models, **pars)
+            step.do_detection()
+
+            state = "COMPLETE"
+
+            if not self.save_intermediate_results:
+                self.log.debug(
+                    "The following files will be deleted since \
+                    save_intermediate_results=False:"
+                )
+            for model in self.input_models:
+                model.meta.cal_step["outlier_detection"] = state
+                if not self.save_intermediate_results:
+                    #  Remove unwanted files
+                    crf_path = self.make_output_path(basepath=model.meta.filename)
+                    # These lines to be used when/if outlier_i2d files follow
+                    # output_dir crf_file = os.path.basename(crf_path)
+                    # outlr_path = crf_path.replace(crf_file, outlr_file)
+                    outlr_file = model.meta.filename.replace("cal", "outlier_i2d")
+                    blot_path = crf_path.replace("crf", "blot")
+                    median_path = blot_path.replace("blot", "median")
+
+                    for fle in [outlr_file, blot_path, median_path]:
+                        if os.path.isfile(fle):
+                            os.remove(fle)
+                            self.log.debug(f"    {fle}")
+
+            return self.input_models
 
         # if input is not valid then log a warning message and
         # skip outlier detection step
         self.log.warning(
-            "Input contains either < 2 elements or the wrong exposure type \
+            "Input is not a ModelContainer, \
+            does not contains >= 2 elements, \
+            or the elements contain the wrong exposure type \
             (i.e. meta.exposure.type != 'WFI_IMAGE')."
         )
-        for model in self.input_models:
-            model.meta.cal_step.outlier_detection = "SKIPPED"
+        if self.input_container:
+            for model in self.input_models:
+                model.meta.cal_step["outlier_detection"] = "SKIPPED"
         self.skip = True
-        return self.input_models
-
-    def _run_outlier_detection(self):
-        # Setup output path naming if associations are involved.
-        asn_id = self.input_models.asn_table.get("asn_id", None)
-        if asn_id is not None:
-            _make_output_path = self.search_attr("_make_output_path", parent_first=True)
-            self._make_output_path = partial(_make_output_path, asn_id=asn_id)
-
-        detection_step = outlier_detection.OutlierDetection
-        pars = {
-            "weight_type": self.weight_type,
-            "pixfrac": self.pixfrac,
-            "kernel": self.kernel,
-            "fillval": self.fillval,
-            "nlow": self.nlow,
-            "nhigh": self.nhigh,
-            "maskpt": self.maskpt,
-            "grow": self.grow,
-            "snr": self.snr,
-            "scale": self.scale,
-            "backg": self.backg,
-            "kernel_size": self.kernel_size,
-            "save_intermediate_results": self.save_intermediate_results,
-            "resample_data": self.resample_data,
-            "good_bits": self.good_bits,
-            "allowed_memory": self.allowed_memory,
-            "in_memory": self.in_memory,
-            "make_output_path": self.make_output_path,
-            "resample_suffix": "i2d",
-        }
-
-        self.log.debug(f"Using {detection_step.__name__} class for outlier_detection")
-
-        # Set up outlier detection, then do detection
-        step = detection_step(self.input_models, **pars)
-        step.do_detection()
-
-        state = "COMPLETE"
-
-        if not self.save_intermediate_results:
-            self.log.debug(
-                "The following files will be deleted since save_intermediate_results=False:"
-            )
-        for model in self.input_models:
-            model.meta.cal_step.outlier_detection = state
-            if not self.save_intermediate_results:
-                #  Remove unwanted files
-                crf_path = self.make_output_path(basepath=model.meta.filename)
-                #  These lines to be used when/if outlier_i2d files follow output_dir
-                #  crf_file = os.path.basename(crf_path)
-                #  outlr_path = crf_path.replace(crf_file, outlr_file)
-                outlr_file = model.meta.filename.replace("cal", "outlier_i2d")
-                blot_path = crf_path.replace("crf", "blot")
-                median_path = blot_path.replace("blot", "median")
-
-                for fle in [outlr_file, blot_path, median_path]:
-                    if os.path.isfile(fle):
-                        os.remove(fle)
-                        self.log.debug(f"    {fle}")
-
         return self.input_models
